# server/app/src/tsoa-build/openapi.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ErrorResponseBody_AuthErrorResponseContent.AuthErrorBodyMessageT_": {
				"description": "Represents the response body of an error response",
				"properties": {
					"errors": {
						"items": {
							"$ref": "#/components/schemas/AuthErrorResponseContent"
						},
						"type": "array"
					},
					"message": {
						"$ref": "#/components/schemas/AuthErrorBodyMessageT"
					}
				},
				"required": [
					"errors",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthErrorResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_AuthErrorResponseContent.AuthErrorBodyMessageT_",
				"description": "The response body for an {@link AuthError}"
			},
...
			"ValidationErrorResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_ValidationErrorResponseContent.ValidationErrorMessageT_",
				"description": "Represents the response body for a validation error\n\nUsed by internal {@link QueryValidationError} and {@link ClassValidationError},\nas well as `sequelize`.{@link ValidationError} and `tsoa`.`ValidateError`"
			},
...
			"OrderByFieldValues": {
				"description": "Possible values for the filter query paramater on controller routes",
				"enum": [
					"createdAt",
					"updatedAt",
					"-createdAt",
					"-updatedAt"
				],
				"type": "string"
			},
...
			"NotFoundResponseBody": {
				"$ref": "#/components/schemas/ErrorResponseBody_NotFoundResponseContent.NotFoundMessageT_",
				"description": "The type of {@link ErrorResponseBody} for not found responses"
			},
...
			"AppModelIdT": {
				"type": "number",
				"format": "double",
				"description": "This is the type of {@link AppModel.id} - used to ensure its correct when used in routers etc."
			},
...
			"DataResponseBody_Array_ReadResponseContent_BlogModelDTO.never.imageURL-or-publishedDate__.Blog.FoundResponseMessageT_": {
				"description": "Represents the response body of a data response",
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReadResponseContent_BlogModelDTO.never.imageURL-or-publishedDate_"
						},
						"type": "array"
					},
					"message": {
						"$ref": "#/components/schemas/FoundResponseMessageT"
					},
					"metadata": {
						"$ref": "#/components/schemas/MetaData_Blog_"
					}
				},
				"required": [
					"data",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BlogReadManyResponse": {
				"$ref": "#/components/schemas/DataResponseBody_Array_ReadResponseContent_BlogModelDTO.never.imageURL-or-publishedDate__.Blog.FoundResponseMessageT_",
				"description": "Possible responses for {@link BlogController.getAllBlogs}"
			},
			"BlogNotFoundResponseBody": {
				"$ref": "#/components/schemas/NotFoundResponseBody",
				"description": "Response type for this controllers object not found response"
			},
...
			"BlogReadResponse": {
				"$ref": "#/components/schemas/DataResponseBody_ReadResponseContent_BlogModelDTO.never.imageURL-or-publishedDate_.Blog.FoundResponseMessageT_",
				"description": "Possible responses for {@link BlogController.getBlog}"
			},
...
			"BlogCreateResponse": {
				"$ref": "#/components/schemas/DataResponseBody_CreateResponseContent_BlogModelDTO.never.imageURL-or-publishedDate_.Blog.CreatedResponseMessageT_",
				"description": "Possible responses for {@link BlogController.createBlog}"
			},
...
			"CreateRequestBody_BlogModelDTO.user-or-tags.imageURL-or-publishedDate_": {
				"$ref": "#/components/schemas/TransformKeys_BlogModelDTO.Extract_user-or-tags-or-CreateRequestBodyOmittedKeysT.keyofBlogModelDTO_.imageURL-or-publishedDate.never_"
			},
			"BlogCreateRequest": {
				"$ref": "#/components/schemas/CreateRequestBody_BlogModelDTO.user-or-tags.imageURL-or-publishedDate_",
				"description": "Request body format for {@link BlogController.createBlog}'s `blog` body parameter"
			},
...
			"BlogUpdateResponse": {
				"$ref": "#/components/schemas/DataResponseBody_UpdateResponseContent_BlogModelDTO.never.imageURL-or-publishedDate_.Blog.UpdatedResponseMessageT_",
				"description": "Possible responses for {@link BlogController.updateBlog}"
			},
...
			"BlogUpdateRequest": {
				"$ref": "#/components/schemas/UpdateRequestBody_BlogModelDTO.user-or-tags.never_",
				"description": "Request body format for {@link BlogController.updateBlog}'s `blog` body parameter"
			},
...
			"BlogDeleteResponse": {
				"$ref": "#/components/schemas/DeleteResponseBody_Blog_",
				"description": "Possible responses for {@link BlogController.deleteBlog}"
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "api_key",
				"in": "query"
			},
			"oauth2": {
				"type": "oauth2",
				"flows": {
					"password": {
						"tokenUrl": "/auth/login",
						"scopes": {
							"admin": "modify things",
							"staff": "modify some things",
							"user": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": ""
	},
	"paths": {
...
		"/blogs": {
			"get": {
				"operationId": "GetAllBlogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogReadManyResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of all blogs, matching the query parameters if provided.",
				"tags": [
					"Blog"
				],
				"security": [],
				"parameters": [
					{
						"description": "- Retrieve records starting from this offset",
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Retrieve up to this many records",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- Order records by this attribute",
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/OrderByFieldValues"
						}
					},
					{
						"description": "- Stringified JSON where the keys represent a property of a blog and values represent a text based filter for the property",
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateBlog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogCreateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Creates a new blog",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"admin",
							"staff"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- The posted body object, containing the creation attributes of BlogModel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BlogCreateRequest",
								"description": "- The posted body object, containing the creation attributes of BlogModel"
							}
						}
					}
				}
			}
		},
		"/blogs/{id}": {
			"get": {
				"operationId": "GetBlog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogReadResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Retrieves the details of an existing blog with a matching id",
				"tags": [
					"Blog"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The identifier of the blog to get",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateBlog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogUpdateResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Update a blog with a matching id",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"admin",
							"staff"
						]
					}
				],
				"parameters": [
					{
						"description": "- The identifier of the blog to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					}
				],
				"requestBody": {
					"description": "- The posted body object, containing some or all modifiable attributes of BlogModel",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BlogUpdateRequest",
								"description": "- The posted body object, containing some or all modifiable attributes of BlogModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteBlog",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogDeleteResponse"
								}
							}
						}
					},
					"401": {
						"description": "Not Authenticated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"403": {
						"description": "Not Authorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthErrorResponseBody"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BlogNotFoundResponseBody"
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponseBody"
								}
							}
						}
					}
				},
				"description": "Delete the blog with the matching id",
				"tags": [
					"Blog"
				],
				"security": [
					{
						"api_key": []
					},
					{
						"oauth2": [
							"admin",
							"staff"
						]
					}
				],
				"parameters": [
					{
						"description": "- The identifier of the blog to delete",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/AppModelIdT"
						}
					},
					{
						"description": "- If true the blog will actually be deleted from the database rather than a soft delete (handled by sequelize)",
						"in": "query",
						"name": "force",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"basePath": "/",
...
	"securityDefinitions": {
		"api_key": {
			"type": "apiKey",
			"name": "api_key",
			"in": "query"
		},
		"oauth2": {
			"type": "oauth2",
			"tokenUrl": "/auth/login",
			"flow": "password",
			"scopes": {
				"admin": "modify things",
				"staff": "modify some things",
				"user": "read things"
			}
		}
	}
}